refactor(ui): extract GeodataMode enum into types
Moves the GeodataMode enum out of App.tsx into src/types/plugin.ts so
it sits next to the other plugin-shell shapes (PluginProps,
WorkbenchResult, GeoConfig). App.tsx, main.tsx and App.spec.tsx now
import it from './types' via the barrel. No behavior change.
